+++ src/components/UserProfile/page/index.tsx
@@ -3,6 +3,8 @@ import {
   editProfileModalAtom,
   loggedAtom,
   logoutModalAtom,
+  MyLikeAtom,
+  MyPostAtom,
 } from "Atoms";
 import { PostBox } from "components/Common";
 import Category from "components/Common/Category";
@@ -14,30 +16,32 @@ import { useQuery } from "react-query";
 import { useNavigate, useParams } from "react-router-dom";
 import { useRecoilState } from "recoil";
 import * as S from "./style";
+import * as I from "Assets/svg";
 import EditProfileModal from "components/Modal/EditProfileAtom";
 import { DEFAULT_PROFILE_IMAGE } from "shared/config";
 import UserProfilePageSkeleton from "../skeleton";
 import LogoutModal from "components/Modal/LogoutModal";
 import { marked } from "marked";
 import Page404 from "components/404";
+import MyPost from "../MyPost";
+import MyLikePost from "../MyPost";
 
 export default function UserPropfile() {
   const [userInfo, setUserInfo] = useState<any>({});
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [profileImg, setProfileImg] = useState<string>(DEFAULT_PROFILE_IMAGE);
-  const [feedList, setFeedList] = useState<any[]>([]);
   const [isMine, setIsMine] = useState<boolean>(false);
   const [postsNull, setPostsNull] = useState<boolean>(false);
   const [is404, setIs404] = useState<boolean>(false);
+  const params = useParams();
 
+  const [myPost, setMyPost] = useRecoilState(MyPostAtom);
+  const [myLike, setMyLike] = useRecoilState(MyLikeAtom);
   const [logged] = useRecoilState(loggedAtom);
   const [logoutModal, setLogoutModal] = useRecoilState(logoutModalAtom);
   const [deletePostModal] = useRecoilState(deletePostModalAtom);
   const [editProfileModal, setEditProfileModal] =
     useRecoilState(editProfileModalAtom);
-
-  const navigator = useNavigate();
-  const params = useParams();
 
   useEffect(() => {
     const getUserInfo = async () => {
@@ -50,9 +54,9 @@ export default function UserPropfile() {
 
         setProfileImg(res.data.profileImageUrl);
         setIsMine(res.data.isMine);
-        setFeedList(res.data.feedList);
         setUserInfo(res.data);
         setIsLoading(false);
+
         if (res.data.feedList.length === 0) {
           setPostsNull(true);
         }
@@ -65,34 +69,15 @@ export default function UserPropfile() {
     getUserInfo();
   }, [params.ninkname]);
 
-  const fetch = async () => {
-    try {
-      const res: any = await user.getUserInfo(
-        JSON.parse(localStorage.getItem("token") || "{}").accessToken,
-        String(params.nickname)
-      );
-      setIsMine(res.data.isMine);
-      setFeedList(res.data.feedList);
-      setUserInfo(res.data);
-    } catch (e: any) {
-      console.log(e);
-    }
+  const clickMyPost = () => {
+    setMyPost(true);
+    setMyLike(false);
   };
 
-  const userProfileQuery = useQuery({
-    queryKey: "UserProfile",
-    queryFn: fetch,
-    refetchOnWindowFocus: false,
-  });
-
-  useEffect(() => {
-    if (isMine) {
-      if (!logged) {
-        navigator("/signin");
-      }
-    }
-    fetch();
-  }, [params.nickname]);
+  const clickMyLike = () => {
+    setMyPost(false);
+    setMyLike(true);
+  };
 
   return (
     <>
@@ -128,36 +113,22 @@ export default function UserPropfile() {
           </S.ProfileLayout>
           <S.MyPostsLayout>
             <S.CategoryBox>
-              <Category>
-                {isMine
-                  ? "💻내 게시물's"
-                  : `💻 ${userInfo.nickname}님의 게시물's`}
-              </Category>
+              {isMine ? (
+                // 마이페이지일때 좋아요 누른 게시물이 보이게하기
+                <S.MyCategoryBox>
+                  <S.MyCategory clicked={myPost} onClick={clickMyPost}>
+                    💻내 게시물's
+                  </S.MyCategory>
+                  <S.MyCategory clicked={myLike} onClick={clickMyLike}>
+                    <I.Like /> 하트
+                  </S.MyCategory>
+                </S.MyCategoryBox>
+              ) : (
+                <Category>{`💻 ${userInfo.nickname}님의 게시물's`}</Category>
+              )}
             </S.CategoryBox>
             {isLoading && <UserProfilePageSkeleton />}
-            {postsNull ? (
-              <PostIsNull />
-            ) : (
-              <S.MyPostsBox>
-                {feedList.map((post) => (
-                  <PostBox
-                    key={post.id}
-                    id={post.id}
-                    title={post.title}
-                    imageUrl={post.thumbnail}
-                    content={
-                      marked(post.previewContent).replace(/<[^>]+>/g, "") +
-                      "..."
-                    }
-                    view={post.hit}
-                    like={post.likeCount}
-                    isMine={isMine}
-                    isDefault={false}
-                    inUserPage={true}
-                  />
-                ))}
-              </S.MyPostsBox>
-            )}
+            {myLike ? <MyLikePost /> : <MyPost />}
           </S.MyPostsLayout>
         </>
       )}
